Add files via upload
ompleted my First task by Prodigy InfoTech

#ProdigyInfoTech

Project #1 for Data Science (ProdigyInfotech)

Q) Create a bar chart or histogram to visualize the distribution of a categorical or continuous variable such as the distribution of ages or genders in a population.

In this, I worked on Google Colab to implement the project. Also, Used drive mount to feature so that the pandas library can directly extract the required csv dataset file from the google drive.